html/semantics/embedded-content/media-elements/ready-states/autoplay.html

<!doctype html>
<title>autoplay</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/media.js"></script>
<div id="log"></div>
<script>
function autoplay_test(tagName, src) {
  async_test(function(t) {
    var e = document.createElement(tagName);
    e.src = src;
    e.autoplay = true;
    var actual_events = [];
    var expected_events = ['canplay', 'play', 'playing', 'canplaythrough'];
    expected_events.forEach(function(type) {
      e.addEventListener(type, t.step_func(function() {
        assert_equals(e.readyState, e.HAVE_ENOUGH_DATA);
        assert_false(e.paused);
        actual_events.push(type);
        if (type == 'canplaythrough') {
          assert_array_equals(actual_events, expected_events);
          t.done();
        }
      }));
    });
  }, tagName + '.autoplay');
}

autoplay_test('audio', getAudioURI('/media/sound_5'));
autoplay_test('video', getVideoURI('/media/movie_5'));
</script>
